backend/requirements.txt
fastapi==0.115.0
uvicorn==0.28.0
sqlalchemy==2.0.38
python-dotenv==1.0.1
pydantic==2.7.0
openai==1.27.0
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pytz==2023.3 